+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to the package will be documented here.
 
 ---
 
+## [4.12.42] - 2025-06-21
+
+- **Updated** JSDoc some functions and methods.
+- **Updated and Optimized** `getDatesInRange()` method in `Chronos`. Added new option `onlyDays` to get dates for only the provided days.
+- **Allowed** `formatStrict()` method in `Chronos` to accept other string values [made less strict].
+
 ## [4.12.41] - 2025-06-17
 
 - **Updated** `getDatesInRange()` and `getDatesForDay()` `Chronos` methods' options to change the date rounding behaviour.